Revert "[graphite] Use SampleCount enum for sample count in TextureInfo"
This reverts commit cd6d01ab68ff8936b902a91d433bd87b742ca160.

Reason for revert: Blocking Android roll

Original change's description:
> [graphite] Use SampleCount enum for sample count in TextureInfo
>
> Bug: b/452351267
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100357
> Commit-Queue: Michael Ludwig <michaelludwig@google.com>

Bug: b/452351267
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1104856
Commit-Queue: Robert Phillips <robertphillips@google.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>

+++ tools/gpu/vk/VkYcbcrSamplerHelper.cpp
@@ -213,7 +213,7 @@ skgpu::VulkanAlloc alloc;
     alloc.fSize = requirements.size;
 
     skgpu::graphite::VulkanTextureInfo imageInfo = {
-            vkImageInfo.samples,
+            static_cast<uint32_t>(vkImageInfo.samples),
             skgpu::Mipmapped::kNo,
             VK_IMAGE_CREATE_PROTECTED_BIT,
             vkImageInfo.format,
